pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbrush"
version = "0.1.0"
description = "Image editing toolkit for AI agents. Photoshop for Claude Code, Codex, Cursor, and friends."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Ultrathink", email = "ceo@ultrathink.art" },
]
keywords = ["image-editing", "ai-agents", "pillow", "background-removal", "greenscreen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "Pillow>=9.0",
]

[project.optional-dependencies]
generate = ["openai>=1.0"]
all = ["openai>=1.0"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.urls]
Homepage = "https://github.com/ultrathink-art/agentbrush"
Repository = "https://github.com/ultrathink-art/agentbrush"
Issues = "https://github.com/ultrathink-art/agentbrush/issues"

[project.scripts]
agentbrush = "agentbrush.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentbrush"]

[tool.pytest.ini_options]
testpaths = ["tests"]
